maint: Fix invalid unref in lp_scene_bus_callback
* lib/lp-scene.c (lp_scene_bus_callback): Unref GstEvent using
gst_event_unref (not g_object_unref).
* lib/lp-event.c: Fix warning.
* examples/dump.c: Dump event to stdout using to_string.

+++ lib/lp-event.c
@@ -113,7 +113,7 @@ _lp_event_to_string (lp_Event *event, const char *fmt, ...)
                          : LP_IS_MEDIA (event->priv->source) ? "lp_Media"
                          : "unknown",
                          event->priv->source,
-                         event->priv->mask,
+                         (unsigned int) event->priv->mask,
                          suffix);
   g_assert_nonnull (str);
   g_free (suffix);
